 healthcare_platform/requirements.txt
uvicorn
fastapi
pydantic
python-dotenv
